Added working integration and system tests
