+++ frontend/cypress/support/commands.js
@@ -1,3 +1,6 @@
+// SPDX-License-Identifier: MIT
+// SPDX-FileCopyrightText: 2021 Dominic Heil <d.heil@campus.tu-berlin.de>
+
 // ***********************************************
 // This example commands.js shows you how to
 // create various custom commands and overwrite
@@ -23,6 +26,16 @@
 //
 // -- This will overwrite an existing command --
 // Cypress.Commands.overwrite('visit', (originalFn, url, options) => { ... })
+
+export const mobileScreenSettings = {
+	viewportHeight: 812,
+	viewportWidth: 375
+};
+
+export const desktopScreenSettings = {
+	viewportHeight: 1080,
+	viewportWidth: 1920
+};
 
 Cypress.Commands.add("setDesktopScreenSize", () => cy.viewport(1920, 1080));
