+++ frontend/package.json
@@ -31,7 +31,7 @@
     "eject": "react-scripts eject",
     "lint": "lint-staged",
     "prepare": "cd .. && husky install",
-    "cy:open": "cypress open-ct",
+    "cy:open": "cypress open-ct --config-file cypress.json",
     "cy:run": "cypress run-ct",
     "cy:open-system": "cypress open --config-file cypress_system_tests.json --env REACT_APP_PROJECT=618583cbf0506,REACT_APP_ENDPOINT=http://localhost:81/v1,REACT_APP_KEY=f9c23feb7e5e70217a40516490d6ed3572c863f27716e183dbbc37d7d10dfb928434acea06a05637bfb7f294652336cb08789866e57ce41f81856bf15bd3c8f0dc646f32ab7fb04f6b437850941d0f9beefd3aa95097f06c6a5cd6e567594a1fc023d529b217ac33cb266124de2647a6bd0666acef34ca8eba4332e56f64a13a",
     "cy:run-system": "cypress run --config-file cypress_system_tests.json"
@@ -57,11 +57,13 @@
   "devDependencies": {
     "@babel/core": "^7.13.0",
     "@cypress/react": "^5.10.3",
+    "@cypress/webpack-dev-server": "^1.7.0",
     "cypress": "^9.1.1",
     "eslint": "^7.32.0",
     "eslint-config-ideal": "^0.1.0",
     "eslint-plugin-cypress": "^2.12.1",
     "eslint-plugin-react": "^7.26.1",
+    "html-webpack-plugin": "^4.5.2",
     "husky": "^7.0.4",
     "lint-staged": "^11.2.6",
     "node-appwrite": "^3.0.0",
